Adaptations for data source other than yahoo
yahoo's historic data service (ichart.yahoo.com) is no longer available.
Implemented two fixes for this: A (partially implemented[1]) mechanism for
retrieving historical data by calling an external program/script.
(Specifically, look at 'perform_external_http_retrieval' and
'external_webrequest' (in HTTP_DATA_RETRIEVAL) for the heart of the
implementation.) The other fix (implemented when I realized that it was
simple [i.e., not time-consuming]) was to change the data "conversion" code
in TRADABLE_DATA_CONVERSION to handle sources other than yahoo -
specifically, quotemedia.com - by lopping off extra fields (i.e., beyond
the standard date,o,h,l,c,vol) at the end of each line/record. Other than
this change, the quotemedia format is basically compatible with the yahoo
format. (To adapt to other sources, more changes may be needed, of course.)
The heart of this implementation is in 'converted_yahoo_line'. A couple new
clusters needed to be added to the Eiffel configuration file (mas-
extended.ecf) for a couple GOBO classes used for this implementation.

The example mas_httprc file (./build/config/rc_files/mas_httprc) was
modified (and tested) to use the new quotemedia source/host. Also, improved
the getdata.rb script (used by the retrieve-via-external - script solution
described above) with some refactoring. More refactoring (and probably a
new class/file or two) will be needed if the "external- retrieval" solution
will be used officially. MAS_PRODUCT_INFO was updated, as well: incremented
the version number and updated the date.

[1] partially implemented in the sense that although it can be made
functional with small and easy code changes, it is not complete in that it
is not a proper (i.e., properly refactored) design and is not configurable
at runtime. For example, the mas_httprc file could take an option to
specify use of this external query method (and perhaps the script to use to
invoke it). If this functionality is put to serious use, then, of course,
the needed missing implementation/design should be completed.