Began: external data retrieval in file-data-mode
Began work on option to, when the MAS server is run in the mode
to get its stock/tradable data from files: if the file for the
current symbol does not exist, retrieve the needed data, via an
external script, from a web site and store that data in the
expected file/filename/path. (Note: a complete implementation
will also require retrieving data when the file exists, but is
out of date.) The first iteration of the implementation
(functional, but incomplete) of this feature adds a query -
external_data_service_active - to FILE_BASED_TRADABLE_LIST that
indicates whether this externdal-data-retreival functionality is
to be invoked. If the query is true and if data is needed (i.e.,
the data file doesn't exist), retrieve_tradable_data is called to
perform the retrieval. This feature is implemented in
FILE_TRADABLE_LIST. In HTTP_LOADING_FILE_TRADABLE_LIST
external_data_service_active is awlays false and thus
retrieve_tradable_data is never called.  This commit includes
the two ruby files created to implement (again, functional, but
not complete) the external retrieval functionality.

Note that similar functionality exists for the http-retrieval
mode (-w option, in which mas will retrieve the data from a web
site based on the host/path specified in mas_httprc) - that is,
retrieving data from a web query via an external script
instead of a direct web query. (See HTTP_DATA_RETRIEVAL and
TRADABLE_DATA_CONVERSION.) And I had started some work on
improving that functionality, which (including messy debug-code
and temporary comments) is likely now obsolete and will probably
be removed in a near-future commit.