Add (not reject) "unknown" requested symbol, ...
To allow adding of a (newly requested) tradable symbol to the
list of symbols rather than rejecting it as before: Changed
'symbols' in TRADABLE_LIST from a LIST to a DYNAMIC_LIST. This
allows a new symbol to be easily added by simply calling
'put_front' to put it at the beginning of a list. File-based
descendants of TRADABLE_LIST also have a 'file_names' list that
needs to be synchronized with 'symbols'. Thus this attribute was
also changed to DYNAMIC_LIST. The "add symbol" functionality is
implemented in FILE_TRADABLE_LIST (with the previously-mentioned
put_front call on the symbol and file lists), in 'add_symbol'.
'remove_symbol' also needed to be added to reverse the "add",
since 'add_symbol' needs to be called before the actual data is
retrieved - if the data retrieval fails, the symbol and filename
lists need to be restored to their previous state with
remove_symbol.  Also, the cache of tradables in TRADABLE_LIST:
    cache: HASH_TABLE [TRADABLE [BASIC_TRADABLE_TUPLE], HASHABLE]
is also affected by adding a new symbol.  To make the hash-related
logic easier with respect to the change, the key of the hash table
was changed from INTEGER to HASHABLE.  This required changing any
variables/features/... used as an index into the cache from INTEGER to
HASHABLE.  The key is now the symbol name (STRING, of course) - it
used to be the current symbol-list index.

Another aspect of the implementation of adding a new
symbol/tradable is the 'attempt_to_add' procedure of
TRADABLE_DISPENSER (deferred - it is implemented in
TRADABLE_LIST_HANDLER). It is called when an "unknown" symbol is
requested by the client and it has a precondition of
'expandable' - a new boolean query added to TRADABLE_DISPENSER.
Currently, this feature is only implemented for file-based data,
so 'expandable' should be false for the other (http, database,
...) data sources (and I believe it is). 'attempt_to_add' calls
'add_tradable_for' in TRADABLE_LIST, which executes the high-
level logic of adding a new symbol/tradable - i.e.:
    add_symbol(symbol)
    setup_input_medium
    if fatal_error then
        remove_symbol(symbol)
    end

HTTP_DATA_RETRIEVAL: Removed old/experimental code for retrieving
data via an external HTTP request to a data-providing web site.

Made scanning of tradable input more robust by using the
"record_error_limit" logic/option added to DATA_SCANNER in the
eiffel_library.  DATA_SCANNER is used for other tasks besides
just scanning tradable data, so this mechanism is only invoked
for tradable scanning by defining record_error_limit in
TRADABLE_TUPLE_DATA_SCANNER with value 100 (rather than 0 in the
other cases).  This fixes the "problem" in which a tradable whose
data file has garbage in the first line causes the scanning
process to produce 0 records - an empty data set.

external data retrieval: "Formalized" the path to the external
script by defining it as 'data_retrieval_script_path' in
APP_ENVIRONMENT. 'retrieve_tradable_data', in
FILE_TRADABLE_LIST, now uses that query instead of the hard-
coded path it used before.

MAS_COMMAND_LINE_UTILITIES: Added a (internal) 'lax_security'
query to allow a test/debug facility: If it is false, a macl (socket-
connected console) client can execute shell commands via '!<cmd>'
syntax. (The output goes to the server output stream, rather than
being sent to the macl client.)

Also: changed retrieve_tradable_config.rb to be a class and made the
corresponding changes to retrieve_tradable_data.rb.  And changed the
filter function (in retrieve_tradable_config.rb) to exclude, for the
first 10 lines of the input, the line if it is garbage (i.e., an
error message or other non-tradable-data).  Changed the URL_Query
class to handle the new requirement implied by this (1st 10 lines...)
change: pass the current_line_index to the filter function.