Failure exit status for ruby scripts...
Modified retrieve_tradable_data to exit with non-0 status if the
data retrieval fails. (This is indicated by the stooq server by
it's returning only one line consisting of: No data) This change
allows the MAS server to respond to the client with error
message/status when it asks for a symbol for a non-existent
stock/tradable, which in turns allows the client (TAT) to discard
the symbol and show an error message rather than adding the symbol
to its list.

MAS_COMMAND_LINE_UTILITIES: Changed the lax_security feature to a
constant with value true - in order to allow shell-command
execution with macl for testing. Added '!!!*' comments to remind
myself to change this back for production.