Revert "param_req - param_list is now sorted."
This reverts commit 476763df6433e6f8388148db5ba9854bab5e129c.
Removed the logic from this commit that changed the parameter
list (server's response to param_req - parameter request) to be
sorted before it is sent to the client.  This change was to
try to ensure that the right parameter was identified by the
client for a parameter modification request.  A better solution
appears to be a redesign to ensure that all parameter names are
always unique, including when the parameters are stored in the
*persist files in "indicators" and (event) "generators".  These
two lines from that commit, for example, are removed:

+    position_of_parameter (p: FUNCTION_PARAMETER): INTEGER

+            Result := sorter.sorted_list(Result)