macl: work on ensuring uniqueness of stored names.
macl: Needed more robust processing of user input for "indicator"
and "market_analyzer" creation: If the user's choice doesn't match
the choices listed by the server, check if the choice uniquely
matches a substring of one of the choices and assume that is the
intended choice.  This is especially (and, perhaps, only) needed for
choosing a "shared" object (e.g., "shared CLOSING_PRICE"), where,
with the new changes that add extra information to object names
to try to ensure they are unique, the name is now different, and
longer, than the corresponding name in the earlier version of the
server.  This messes up the macl scripts (such as "indicators" and
"market_analyzers"), which use these old names.  With this change
creating indicators and market analyzers with these two files now
works again.  The change still needs to be tested with the other
indicator/analyzer files (e.g., ema_of_slope_of_ema, parabolic-sar,
etc.)
macl.ecf: This file hadn't been used for compilation for a few years,
apparently.  Some changes were needed to allow compiling macl with
Eiffel_17.01.