function-parameter-name work re. "parent"s
Began a new strategy for ensuring that function-parameter names are
unique - using the "parent", "initialize_from_parent", etc.
functionality newly added to TREE_NODE and descendants (COMMANDs,
etc.) from the eiffel_library.  The main work for this iteration is
to initialize an object's parent at the right place - e.g., when the
parent is being constructed and/or when the parent's "set_operand*"
procedure is called.  The goal, of course, is to have all parents
correctly set when a tree (e.g., indicator or analyzzer with its
contained "functions" and "commands/operators") has been completely
created/initialized.

Also, added some "append_to_name" calls at appropriate places.
(The append_to_name functionality might or might not be used
in the final implementation of parameter-name uniqueness. If it
is not, it will probably be removed.)

Also, added a full_name query to TRADABLE_FUNCTION and
COMPLEX_FUNCTION (similar to that in COMMAND), which might or
might not be used for the parameter-name implementation.
(Again, if not, it might be removed.)