function-param-name: parent+s, ancestors, etc.
Completed the main implementation work on the solution to ensuring
that all parameter names, as presented to the client (for display
and editing), for an indicator or analyzer (event-generator) are
unique: Changes corresponding to the change in TREE_NODE of having
multiple parents instead of just one parent, and changing the
implementation of 'ancestors' to a tree instead of a list.  For
example, FUNCTION_PARAMETER has new 'set_name' and 'owner' features.
'owner' is used in function 'summary' of SESSION_FUNCTION_PARAMETER
to obtain the ancestors (parents, parents' parents, ...) of the
"source_parameter" (the parameter that is actually attached to the
target function or command object, which is, i.e., the "owner"):
             ancs := param.owner.ancestors
to produce an alternate name (based on a "summary") for the
parameter, based on its "owner".  'unique_param_name' calls 'summary'
to attempt to create a new name for the parameter that does not
conflict with another parameter's name.  The 'name' function calls
'unique_param_name' when it finds that the target parameter's name
(source_parameter.unique_name query [which probably needs a name
change itself]) is a duplicate of another parameter name for the
target indicator/analyzer.  (MAS_SESSION keeps track of such name
conflicts via a HASH_TABLE.)  These changes, of course, required
corresponding changes in descendant (of TREE_NODE and of
FUNCTION_PARAMETER) classes.

Also, the parent => parents change in TREE_NODE required changes
in descendants such as: parent_implementation := p ->
parents_implementation.extend(p).

Finally, some changes were made in descendants related to the new
'verbose_name' (in TREE_NODE and FUNCTION_PARAMETER) and
'who_am_i__parent' (in TREE_NODE) features intended as an early
implementation of the unique-name functionality.  These features
will likely go away in an upcoming commit, since they now appear
to be obsolete.