cleanup - code/comment/... removal, etc.
Removed these comments (more than 40):
--- !!!! indexing once_status: global??!!!
They were interspersed throughout the code as a reminder that
once routines may need a status change if the application is made
multi-threaded.  This might happen in the future, but is unlikely.

Removed 'append_to_name' and the 'configure_...' routines that
call it.  This was part of the first attempt at making sure the
parameter names for a particular indicator or analyzer did not
conflict.  With the new solution (using parents/ancestors), this
code is not obsolete/not-needed.

Removed several reminder and/or debugging-related comments, as well
as print/log_error statements, that are no longer needed.

FUNCTION_PARAMETER and descendants: renamed unique_name to
verbose_name.

AGENT_BASED_FUNCTION: implemented calling of initialize_from_parent
on each member of "inputs" to tell these input functions
(TRADABLE_FUNCTIONs) that Current (the AGENT_BASED_FUNCTION) is
their parent.