Adding period-type for event-generation, step 1
Began adding the period-type to event-generator - parameter- and
data-request- commands - because event-generators have a period-
type that they apply to the functions that they process (e.g.,
a "weekly" event-generator will process weekly data) and the
client needs to be able to specify the period-type for event-
data requests, and for event-generator-parameter-change requests,
to make use of this flexibility.  The "first step" was to add a
period-type to the client's request to modify event-generator
parameters - i.e., in MODIFY_EVENT_GENERATOR_PARAMETERS_REQUEST_CMD
(the 'event_param_set_req' in the server protocol), with suitable
hook routines in the parent, PARAMETER_BASED_REQUEST_CMD.  The
other needed changes - adding the period-type to the event-
generator parameter settings request (evgen_param_req) and to the
event-generator-data request (event_data_req) have not yet been
implemented, although the 'server_protocol_grammar' file has
been changed to reflect the intended design.

Also, only the parsing of the newly added period-type from the client-
request (including adjusting to the extra period-type "argument") has
been implemented, not the actual setting of the event-generator
parameters based on period type.  Therefore, although the system
compiles and runs it, of course, runs incorrectly with respect to event-
generation.